plat/kvm: arm64: Build L3 page table for range greater than 2MB
Current L3 table build function only supports to build L3 table for
memory range not exceeded than 2MB. But as we mentioned in previous
patch, the image size will be greater than 2MB. So in this case, build
L3 table for bigger image is not convenient. We have improve the
function of building L3 table to support bigger memory range.

Change-Id: I2dc0b11eb7115bbd3bc33f224bb4315331021633
Jira: ENTOS-825
Signed-off-by: Wei Chen <wei.chen@arm.com>
Signed-off-by: Jia He <justin.he@arm.com>
Reviewed-by: Sharan Santhanam <sharan.santhanam@neclab.eu>