plat: Clean up kernel image symbols
From: Wei Chen <wei.chen@arm.com>
There is a potential undefined behaviour for pointer comparision.
How to avoid this gcc optimization is under discussing. Since it
is not a block issue. I will follow it up in the future.
Provide macro definitions for text,bss,data... is the precondition
for future solution.

Change-Id: I1b9c5a1a88a96fef93f87a63b5f4064b939ef1b1
Jira: ENTOS-655
Signed-off-by: Wei Chen <wei.chen@arm.com>
Signed-off-by: Jia He <justin.he@arm.com>
Reviewed-by: Sharan Santhanam <sharan.santhanam@neclab.eu>