plat/kvm: arm64: Calculate pagetable size dynamically by image size
From: Wei Chen <wei.chen@arm.com>
The previous L3 page table size is a fixed value based on the
assumption that, image size will not exceed 2MB. But that's not
true, some applications like node.js will be about 40MB after
converting to Unikernel. So in this patch, we calculate the
page table size by real image size. In order to mapping the
image occupied memory to L3 page table, we round up the image
size with 2MB. If we don't round up the image size with 2MB.
Some memory that is not occupied by image but shared the same
2MB block with image tail will be unmapped in page table.

Change-Id: I06eb1c4c37691986160c0f9ed230874f7dfb5fa3
Jira: ENTOS-825
Signed-off-by: Wei Chen <wei.chen@arm.com>
Signed-off-by: Jia He <justin.he@arm.com>`