plat/kvm: arm64: Ensure there is no stall data in cache when boot
From: Wei Chen <wei.chen@arm.com>
We will disable MMU and CACHE before pagetable is ready. This
means we will change memory with cache disabled, so we need to
invalidate the cache to ensure there is no stall data in it.
But we don't know the size of the RAM either. And it would be
expensive to invalidate the whole cache. In this case, just
just need to invalidate what we are going to use:
DTB, TEXT, DATA, BSS, pagetables and bootstack.

Change-Id: Ic8b088cef1f81e6ca7dc4d80d2827fa1c251fda8
Jira: ENTOS-655
Signed-off-by: Wei Chen <wei.chen@arm.com>
Signed-off-by: Jia He <justin.he@arm.com>
Reviewed-by: Sharan Santhanam <sharan.santhanam@neclab.eu>
Reviewed-by: Julien Grall <julien.grall@arm.com>